pyproject.toml
[tool.poetry]
name = "cryptowallet"
version = "0.1.0"
description = ""
authors = ["spogins <spogins@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.100.1"
uvicorn = "^0.23.2"
dependency-injector = "^4.41.0"
sqlalchemy = "^2.0.19"
psycopg2 = "^2.9.6"
passlib = "^1.7.4"
email-validator = "^2.0.0.post2"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
pydantic = "^2.1.1"
responses = "^0.23.3"
fastapi-mail = "^1.4.1"
celery = {extras = ["redis"], version = "^5.3.1"}
redis = "^4.6.0"
aioredis = "2.0.0"
propan = {extras = ["async-rabbit"], version = "^0.1.5.17"}
broker = "^0.3.3"
rabbitmq = "^0.2.0"
asgiref = "^3.7.2"
utils = "^1.0.1"
w3 = "^0.0.0"
web3 = "^6.8.0"
eth-account = "^0.9.0"
requests = "^2.31.0"
asyncpg = "^0.28.0"
httpx = "^0.24.1"
python-socketio = "^5.8.0"
alembic = "^1.11.2"
sqlalchemy-utils = "^0.41.1"
boto3 = "^1.28.28"
django-environ = "^0.10.0"


[tool.poetry.group.dev.dependencies]
setuptools = "^68.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
